Fix #98. Relative symlinks can now be tracked properly

+++ lib/commands/track.sh
@@ -7,7 +7,8 @@ function track {
 	if [[ $filename != $HOME/* ]]; then
 		err $EX_ERR "The file $filename must be in your home directory."
 	fi
-	if [[ ! -e $filename ]]; then
+	# If the file is a dead symlink, we track it anyhow, hence the '! -L'
+	if [[ ! -e $filename && ! -L $filename ]]; then
 		err $EX_ERR "The file $filename does not exist."
 	fi
 	home_exists 'track' "$castle"
@@ -52,7 +53,44 @@ function track {
 		fi
 		local remote_folder=$(dirname "$remote")
 		mkdir -p "$remote_folder"
-		mv -f "$local" "$remote"
+
+		# Check if the file is a relative symlink, if so we don't move it but create
+		# an appropriate relative symlink instead that matches the new location
+		if [[ -L $local ]]; then
+			local target=$(readlink "$local")
+			if [[ $target =~ ^/ ]]; then
+				# It's an absolute symlink, just move it
+				mv -f "$local" "$remote"
+			else
+				# Figure out the relative path from the symlink location in
+				# the castle to the path the symlink points at
+
+				# Get the paths to $remote and $local relative to the homedir
+				local path_to_local=$(dirname ${local/#$HOME\//})
+				local path_to_remote=$(dirname ${remote/#$HOME\//})
+				# Replace every part of the path with a '..' (the '[^/]+$' is for the last part)
+				local path_from_remote=''
+				while [[ $path_to_remote =~ ^(.*)[^/]+/+(.*)$ || $path_to_remote =~ ^[^/]+$ ]]; do
+					path_to_remote=${BASH_REMATCH[1]}${BASH_REMATCH[2]}
+					path_from_remote="../${path_from_remote}"
+				done
+				local relpath="${path_from_remote}${path_to_local}/${target}"
+				# Remove unnecessary path parts
+				# Remove '/./'
+				relpath=${relpath//\/.\//\/}
+				# Remove 'somedir/../'
+				while [[ $relpath =~ ^(.*/)?[^/.]+/\.\./(.*)$ ]]; do
+					relpath=${BASH_REMATCH[1]}${BASH_REMATCH[2]}
+				done
+				ln -s $relpath $remote
+				# Remove $remote so we can create the symlink further down
+				rm $local
+			fi
+		else
+			# Just a regular old file. Move it
+			mv -f "$local" "$remote"
+		fi
+		# Create the symlink in place of the moved file (simulate what the link command does)
 		ln -s "$remote" "$local"
 
 		local git_out
